fix warning 4018

+++ cocos/2d/CCLayer.cpp
@@ -1054,7 +1054,7 @@ void LayerMultiplex::switchTo(int n)
 
 void LayerMultiplex::switchTo(int n, bool cleanup)
 {
-    CCASSERT( n < _layers.size(), "Invalid index in MultiplexLayer switchTo message" );
+    CCASSERT(static_cast<size_t>(n) < _layers.size(), "Invalid index in MultiplexLayer switchTo message" );
     
     this->removeChild(_layers.at(_enabledLayer), cleanup);
     
@@ -1065,7 +1065,7 @@ void LayerMultiplex::switchTo(int n, bool cleanup)
 
 void LayerMultiplex::switchToAndReleaseMe(int n)
 {
-    CCASSERT( n < _layers.size(), "Invalid index in MultiplexLayer switchTo message" );
+    CCASSERT(static_cast<size_t>(n) < _layers.size(), "Invalid index in MultiplexLayer switchTo message" );
 
     this->removeChild(_layers.at(_enabledLayer), true);
 #if CC_ENABLE_GC_FOR_NATIVE_OBJECTS
