+++ tests/cpp-tests/Classes/FileUtilsTest/FileUtilsTest.cpp
@@ -1323,12 +1323,12 @@ void TestListFiles::onEnter()
     char cntBuffer[200] = { 0 };
     snprintf(cntBuffer, 200, "'fonts/' %zu, $defaultResourceRootPath %zu",listFonts.size(), list.size());
 
-    for(int i=0;i<listFonts.size();i++)
+    for(size_t i=0;i<listFonts.size();i++)
     {
         CCLOG("fonts/ %d: \t %s", i, listFonts[i].c_str());
     }
 
-    for(int i=0;i<list.size();i++)
+    for(size_t i=0;i<list.size();i++)
     {
         CCLOG("defResRootPath %d: \t %s", i, list[i].c_str());
     }
